refactored go_ethereum tests to support all providers
