+++ tox.ini
@@ -28,7 +28,8 @@ usedevelop=True
 commands=
     core: py.test {posargs:tests/core}
     ens: py.test {posargs:tests/ens}
-    integration-goethereum: py.test {posargs:tests/integration/test_goethereum.py}
+    integration-goethereum-ipc: py.test {posargs:tests/integration/go_ethereum/test_goethereum_ipc.py}
+    integration-goethereum-http: py.test {posargs:tests/integration/go_ethereum/test_goethereum_http.py}
     integration-ethtester: py.test {posargs:tests/integration/test_ethereum_tester.py}
     integration-parity-ipc: py.test -s {posargs:tests/integration/parity/test_parity_ipc.py}
     integration-parity-http: py.test -s {posargs:tests/integration/parity/test_parity_http.py}
